% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binr-package.R
\docType{package}
\name{binr-package}
\alias{binr}
\alias{binr-package}
\title{binr: Cut Numeric Values into Evenly Distributed Groups}
\description{
Package binr (pronounced as "binner") provides algorithms for cutting numerical values exhibiting a potentially highly skewed distribution into evenly distributed groups (bins). This functionality can be applied for binning discrete values, such as counts, as well as for discretization of continuous values, for example, during generation of features used in machine learning algorithms.
}
\section{Copyright}{
 Copyright (C) Collective, Inc.; with portions Copyright (C) Jabiru Ventures LLC
}

\section{License}{
 Apache License, Version 2.0,
available at http://www.apache.org/licenses/LICENSE-2.0
}

\section{Installation from GitHub}{

\code{devtools::install_github("jabiru/binr")}
}

\seealso{
\code{\link{bins}}, \code{\link{bins.quantiles}}, \code{\link{bins.optimize}}, \code{\link{bins.greedy}}
}
\author{
\strong{Maintainer}: Sergei Izrailev \email{sizrailev@jabiruventures.com}

}
\keyword{binning}
\keyword{bins}
\keyword{discretization}
\keyword{features}
\keyword{internal}
\keyword{learning}
\keyword{machine}
