% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binR_manual.r
\name{binR_manual}
\alias{binR_manual}
\title{binR}
\usage{
binR_manual(fx, data, breaks = NULL)
}
\description{
Bin a dataset using a formula and a breaks function
}
\details{
Bin or dkiscretize a dataset using a formula.
}
\author{
Vilhelm von Ehrenheim
}

