% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binR_quantile.r
\name{binR_quantile}
\alias{binR_quantile}
\title{binR}
\usage{
binR_quantile(fx, data, probs = seq(0, 1, 0.25), cumulative = F)
}
\description{
Bin a dataset using a formula and a breaks function
}
\details{
Bin or dkiscretize a dataset using a formula.
}
\author{
Vilhelm von Ehrenheim
}

