% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binR_rpart.r
\name{binR_rpart}
\alias{binR_rpart}
\title{binR}
\usage{
binR_rpart(fx, data, cumulative = F)
}
\description{
Bin a dataset using a formula and a breaks function
}
\details{
Bin or dkiscretize a dataset using a formula.
}
\author{
Vilhelm von Ehrenheim
}

