% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{bins.merr}
\alias{bins.merr}
\title{Binning Error Function}
\usage{
bins.merr(binct, target.bins)
}
\arguments{
\item{binct}{The number of points falling into the bins.}

\item{target.bins}{Number of bins desired; this is also the max number of bins.}
}
\value{
The mean squared error.
}
\description{
Partitioning the data into bins using splitting, merging
and moving optimizes this error function, which is the mean squared error
of point counts in the bins relative to the optimal number of points per bin.
}
\seealso{
\code{\link{bins}}
}
