% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.binR.r
\name{predict.binR}
\alias{predict.binR}
\title{Predict binR}
\usage{
\method{predict}{binR}(obj, newdata)
}
\arguments{
\item{obj}{binR object}

\item{newdata}{the new data.frame to bin}
}
\value{
a data.frame with the binned objects
}
\description{
Apply binning defined in binR object to new dataset
}

