% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart_breaks.r
\name{rpart_breaks}
\alias{rpart_breaks}
\title{rpart Breaks}
\usage{
rpart_breaks(fx, data)
}
\arguments{
\item{fx}{formula to use for computing breaks}

\item{data}{the dataset to use to compute braks}
}
\description{
Compute breaks for binning using rpart
}
\author{
Vilhelm von Ehrenheim
}

