% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{binr}
\alias{binr}
\alias{binr-package}
\title{Cut Numeric Values Into Evenly Distributed Groups (bins).}
\description{
Package \code{binr} (pronounced as "binner") provides algorithms for cutting
numerical values exhibiting a potentially highly skewed distribution into
evenly distributed groups (bins). This functionality can be applied for
binning discrete values, such as counts, as well as for discretization of
continuous values, for example, during generation of features used in machine
learning algorithms.
}
\section{Maintainer}{
 Sergei Izrailev
}

\section{Copyright}{
 Copyright (C) Collective, Inc.
}

\section{License}{
 Apache License, Version 2.0,
   available at http://www.apache.org/licenses/LICENSE-2.0
}

\section{URL}{
 http://github.com/collectivemedia/binr
}

\section{Installation from github}{

\code{devtools::install_github("collectivemedia/binr")}
}
\author{
Sergei Izrailev
}
\seealso{
\code{\link{bins}}, \code{\link{bins.quantiles}}, \code{\link{bins.optimize}}, \code{\link{bins.greedy}}
}
\keyword{64-bit}
\keyword{bigint}
\keyword{csv}
\keyword{delimited}
\keyword{file}
\keyword{integer64}
\keyword{read.csv}

