% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binR.r
\name{binR}
\alias{binR}
\title{binR}
\usage{
binR(fx, data, algorithm = c("quantile", "rpart", "manual"), ...)
}
\description{
Bin a dataset using a formula and a breaks function
}
\details{
Discretize a dataset using a formula.
}
\author{
Vilhelm von Ehrenheim
}

